requirements.txt
absl-py==2.1.0
addict==2.4.0
anyio==3.7.1
appnope==0.1.4
argon2-cffi==21.1.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
Bottleneck==1.3.5
branca==0.6.0
Brotli==1.0.9
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.2.0
cryptography==41.0.7
cycler==0.11.0
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
drjit==0.4.4
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
fiona==1.9.5
Flask==3.0.2
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.25.0
gast==0.5.4
# GDAL==3.6.2
geopandas==0.14.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipydatawidgets==4.3.2
ipykernel==6.29.2
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.2.0
json5==0.9.14
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter_core==5.5.0
jupyter-server==1.24.0
jupyterlab==3.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab-widgets==3.0.5
keras==2.15.0
kiwisolver==1.4.4
libclang==16.0.6
llvmlite==0.42.0
mapclassify==2.5.0
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
mitsuba==3.5.0
ml-dtypes==0.2.0
munkres==1.1.4
nbclassic==0.5.6
nbclient==0.8.0
nbconvert==7.16.0
nbformat==5.7.0
nest_asyncio==1.6.0
networkx==3.1
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.3
oauthlib==3.2.2
open3d==0.17.0
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.18.0
prometheus-client==0.19.0
prompt-toolkit==3.0.42
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyproj==3.6.1
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
pythreejs==2.4.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rpds-py==0.10.6
rsa==4.9
Rtree==1.0.1
scikit-learn==1.3.0
scipy==1.11.4
Send2Trash==1.8.2
setuptools==68.2.2
shapely==2.0.1
sionna==0.16.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==2.2.0
tinycss2==1.2.1
tornado==6.1
tqdm==4.66.1
traitlets==5.14.1
traittypes==0.2.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.41.2
widgetsnbextension==4.0.9
wrapt==1.14.1
xyzservices==2022.9.0
zipp==3.17.0
